refactor: isolate Dropzone stories
