+++ packages/react-aria-components/stories/index.stories.tsx
@@ -11,201 +11,14 @@
  */
 
 import {action} from '@storybook/addon-actions';
-import {Button, Checkbox, Dialog, DialogTrigger, DropZone, FileTrigger, Input, Label, Link, Modal, ModalOverlay, Radio, RadioGroup, SearchField, Switch, Text, TextField, ToggleButton, Toolbar} from 'react-aria-components';
+import {Button, Checkbox, Dialog, DialogTrigger, FileTrigger, Input, Label, Link, Modal, ModalOverlay, Radio, RadioGroup, SearchField, Switch, TextField, ToggleButton, Toolbar} from 'react-aria-components';
 import {classNames} from '@react-spectrum/utils';
-import {FocusRing, mergeProps, useButton, useClipboard, useDrag} from 'react-aria';
-import React, {useRef} from 'react';
+import React from 'react';
 import styles from '../example/index.css';
 
 export default {
   title: 'React Aria Components'
 };
-
-function Draggable() {
-  let buttonRef = useRef(null);
-  let {dragProps, isDragging} = useDrag({
-    getItems() {
-      return [{
-        'text/plain': 'hello world'}];
-    }
-  });
-  let {buttonProps} = useButton({elementType: 'div'}, buttonRef);
-
-  return (
-    <FocusRing focusRingClass={classNames(styles, 'focus-ring')}>
-      <div
-        {...mergeProps(buttonProps, dragProps)}
-        ref={buttonRef}
-        className={classNames(styles, 'draggable', {['dragging']: isDragging})}>
-        Drag me
-      </div>
-    </FocusRing>
-  );
-}
-
-function Copyable() {
-  let {clipboardProps} = useClipboard({
-    getItems() {
-      return [{
-        'text/plain': 'hello world'
-      }];
-    }
-  });
-
-  return (
-    <FocusRing focusRingClass={classNames(styles, 'focus-ring')}>
-      <div
-        {...clipboardProps}
-        role="textbox"
-        aria-label="copyable element"
-        tabIndex={0}
-        className={styles.copyable}>
-        Copy me
-      </div>
-    </FocusRing>
-  );
-}
-
-export const DropzoneExampleWithFileTriggerLink = (props) => (
-  <div>
-    <DropZone
-      {...props}
-      aria-label={'testing aria-label'}
-      className={styles.dropzone}
-      data-testid="drop-zone-example-with-file-trigger-link"
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')}>
-      <FileTrigger onSelect={action('onSelect')}>
-        <Link>Upload</Link>
-      </FileTrigger>
-    </DropZone>
-  </div>
-);
-
-export const DropzoneExampleWithFileTriggerButton = (props) => (
-  <div>
-    <DropZone
-      {...props}
-      className={styles.dropzone}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')}>
-      <FileTrigger onSelect={action('onSelect')} >
-        <Button>Upload</Button>
-      </FileTrigger>
-    </DropZone>
-  </div>
-);
-
-export const DropzoneExampleWithDraggableAndFileTrigger = (props) => (
-  <div>
-    <Draggable />
-    <DropZone
-      {...props}
-      className={styles.dropzone}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')}>
-      <FileTrigger onSelect={action('onSelect')} >
-        <Button>Browse</Button>
-      </FileTrigger>
-      Or drag into here
-    </DropZone>
-  </div>
-);
-
-export const DropZoneOnlyAcceptPNGWithFileTrigger = (props) => (
-  <div>
-    <DropZone
-      {...props}
-      getDropOperation={(types) =>  types.has('image/png') ? 'copy' : 'cancel'}
-      className={styles.dropzone}
-      onPress={action('OnPress')}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')} >
-      <FileTrigger onSelect={action('onSelect')} acceptedFileTypes={['image/png']}>
-        <Button>Upload</Button>
-      </FileTrigger>
-    </DropZone>
-  </div>
-);
-
-export const DropZoneWithCaptureMobileOnly = (props) => (
-  <div>
-    <DropZone
-      {...props}
-      getDropOperation={(types) =>  types.has('image/png') ? 'copy' : 'cancel'}
-      className={styles.dropzone}
-      onPress={action('OnPress')}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')} >
-      <FileTrigger onSelect={action('onSelect')} defaultCamera="environment">
-        <Button>Upload</Button>
-      </FileTrigger>
-    </DropZone>
-  </div>
-);
-
-export const DropzoneExampleWithDraggableObject = (props) => (
-  <div>
-    <Draggable />
-    <DropZone
-      {...props}
-      className={styles.dropzone}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')} >
-      <Text slot="label">
-        DropZone Area
-      </Text>
-    </DropZone>
-  </div>
-);
-
-export const DropzoneExampleWithCopyableObject = (props) => (
-  <div>
-    <Copyable />
-    <DropZone
-      {...props}
-      className={styles.dropzone}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')}>
-      <Text slot="label">
-        DropZone Area
-      </Text>
-    </DropZone>
-  </div>
-);
-
-export const DropzoneWithRenderProps = (props) => (
-  <div>
-    <Draggable />
-    <Copyable />
-    <DropZone
-      {...props}
-      className={styles.dropzone}
-      onPress={action('OnPress')}
-      onDrop={action('OnDrop')}
-      onDropEnter={action('OnDropEnter')}
-      onDropExit={action('OnDropExit')}>
-      {({isHovered, isFocused, isFocusVisible, isDropTarget}) => (
-        <div>
-          <Text slot="label">
-            DropzoneArea
-          </Text>
-          <div>isHovered: {isHovered ? 'true' : 'false'}</div>
-          <div>isFocused: {isFocused ? 'true' : 'false'}</div>
-          <div>isFocusVisible: {isFocusVisible ? 'true' : 'false'}</div>
-          <div>isDropTarget: {isDropTarget ? 'true' : 'false'}</div>
-        </div>
-      )}
-    </DropZone>
-  </div>
-);
 
 export const FileTriggerButton = (props) => (
   <FileTrigger
